requirements.txt
aiofiles==24.1.0
aiohappyeyeballs==2.6.1
aiohttp==3.13.2
aiosignal==1.4.0
alibabacloud-credentials-api==1.0.0
alibabacloud-dysmsapi20170525==3.1.2
alibabacloud-ocr-api20210707==3.1.2
alibabacloud-tea==0.4.3
alibabacloud_credentials==0.3.6
alibabacloud_endpoint_util==0.0.3
alibabacloud_gateway_spi==0.0.3
alibabacloud_openapi_util==0.2.2
alibabacloud_tea_openapi==0.3.13
alibabacloud_tea_util==0.3.13
alibabacloud_tea_xml==0.0.2
APScheduler==3.11.1
asgiref==3.11.0
asttokens==3.0.1
attrs==25.4.0
blinker==1.9.0
certifi==2025.11.12
cffi==2.0.0
charset-normalizer==3.4.4
click==8.3.1
cryptography==46.0.3
darabonba-core==1.0.4
decorator==5.2.1
Django==5.2.9
django-environ==0.12.0
django-simpleui==2025.6.24
djangorestframework==3.16.1
executing==2.2.1
Flask==2.3.3
Flask-Cors==4.0.0
frozenlist==1.8.0
greenlet==3.3.0
idna==3.11
ipython==9.7.0
ipython_pygments_lexers==1.1.1
itsdangerous==2.2.0
jedi==0.19.2
Jinja2==3.1.6
MarkupSafe==3.0.3
matplotlib-inline==0.2.1
multidict==6.7.0
mysqlclient==2.2.7
numpy==2.3.5
pandas==2.3.3
parso==0.8.5
pexpect==4.9.0
prompt_toolkit==3.0.52
propcache==0.4.1
ptyprocess==0.7.0
pure_eval==0.2.3
pycparser==2.23
pycryptodomex==3.23.0
Pygments==2.19.2
PyMySQL==1.1.2
pyOpenSSL==25.3.0
python-alipay-sdk==3.4.0
python-dateutil==2.9.0.post0
pytz==2025.2
redis==7.1.0
requests==2.32.5
six==1.17.0
SQLAlchemy==2.0.44
sqlparse==0.5.4
stack-data==0.6.3
traitlets==5.14.3
typing_extensions==4.15.0
tzdata==2025.2
tzlocal==5.3.1
urllib3==2.6.1
uuid==1.30
wcwidth==0.2.14
Werkzeug==3.1.4
yarl==1.22.0
